requirements.txt
jupyter>=1.0.0
numpy~=1.19.1
scipy~=1.5.0
matplotlib~=3.3.1
sympy~=1.6.2
pyserial~=3.4
seaborn>=0.11.0

scikit-image~=0.17.2